+++ scripts/repo.js
@@ -7,65 +7,9 @@ import path from 'path';
 import unzip from 'unzip';
 import B from 'bluebird';
 import Handlebars from 'handlebars';
-import { quote } from './encode';
-import jQuery from 'jquery';
-import { jsdom } from 'jsdom';
+import { fencedCodeTabifyDocument } from './tabs';
 
-const $ = jQuery(jsdom().defaultView);
 const LANGUAGES = ['en', 'cn'];
-
-function stripLanguageComment (html) {
-  return html.replace(/(<code [^>]*>)(\/\/ [^\s]*)/, '$1');
-}
-
-function capitalize (languageName) {
-  if (languageName === 'php') {
-    return 'PHP';
-  } 
-
-  return languageName.charAt(0).toUpperCase() + languageName.slice(1);
-}
-
-function appendLanguageBlock (tabEl, preEl, languageName, id, active) {
-  const navTag = tabEl.find('.nav');
-  navTag.append(`<li role="presentation" ${active ? 'class="active"' : ''} role="tab">
-    <a href='#${id}' data-toggle='tab'>${capitalize(languageName)}</a>
-  </li>`);
-  const tabPanel = tabEl.find('.tab-content');
-  let html = preEl[0].outerHTML;
-  html = stripLanguageComment(html);
-  tabPanel.append(`<div role="tabpanel" class="tab-pane ${active ? 'active' : ''}" id="${id}">${html.trim()}</div>`);
-}
-
-export function fencedCodeTabify (html) {
-  const jqHTML = $(html);
-  let tabTagHTML = `<div>
-    <ul class="nav nav-tabs" role="tablist">
-    </ul>
-    <div class="tab-content">
-    </div>
-  </div>`;
-
-  jqHTML.find("pre > code").each((index, codeTag) => {
-    const preTag = $(codeTag).parent();
-    const siblings = preTag.nextAll();
-    const tabEl = $(tabTagHTML);
-    let language = capitalize($(codeTag).attr('class'));
-    let tagIndex = 1;
-    appendLanguageBlock(tabEl, preTag, language, `${tagIndex++}_${index}`, true);
-    siblings.each(function (index, siblingEl) {
-      language = $(siblingEl).find('code').attr('class');
-      if (!language) {
-        return false;
-      }
-      appendLanguageBlock(tabEl, $(siblingEl), language, `${tagIndex}_${index}`);
-      $(siblingEl).remove();
-    });
-    preTag.replaceWith(tabEl);
-  });
-
-  return jqHTML.html();
-}
 
 async function unzipStream (readstream, pathToUnzipped) {
   return new B((resolve, reject) => {
@@ -129,13 +73,24 @@ async function alterDocs (pathToDocs) {
     } else {
       const filePath = path.resolve(pathToDocs, file);
 
-      // Encode UTF8 characters
-      const fileContents = await fs.readFile(filePath, 'utf8');
-      await fs.writeFile(filePath, quote(fileContents));
-
       // Rename README.md to index.md
       if (file.toLowerCase() === 'readme.md') {
         await(fs.mv(filePath, path.resolve(pathToDocs, 'index.md')));
+      }
+    }
+  }
+}
+
+async function applyTabs (pathToHTML) {
+  for (let file of await fs.readdir(pathToHTML)) {
+    const stat = await fs.stat(path.resolve(pathToHTML, file));
+    if (stat.isDirectory()) {
+      applyTabs(path.resolve(pathToHTML, file));
+    } else {
+      const filePath = path.resolve(pathToHTML, file);
+      if (path.extname(file) === '.html') {
+        let treatedHTML = fencedCodeTabifyDocument(await fs.readFile(filePath, 'utf8'));
+        await fs.writeFile(filePath, treatedHTML);
       }
     }
   }
@@ -151,17 +106,11 @@ async function buildDocs (pathToDocs) {
     await exec('mkdocs', ['build', '--site-dir', pathToBuildDocsTo], {
       cwd: pathToDocs,
     });
-
-    // Copy to _site (TODO: How do we make Jekyll do this automatically?)
-    /*const docsSubdirectory = path.resolve(__dirname, '..', '_site', 'docs', language);
-    await mkdirp(path.resolve(__dirname, docsSubdirectory));
-    await fs.copyFile(pathToBuildDocsTo, docsSubdirectory);*/
+    await applyTabs(pathToBuildDocsTo);
   }
 };
 
-if (!process.env.TEST) {
-  (async () => {
-    const pathToRepoDocs = await getRepoDocs('appium', 'appium');
-    await buildDocs(pathToRepoDocs);
-  })();
-}
+(async () => {
+  const pathToRepoDocs = await getRepoDocs('appium', 'appium');
+  await buildDocs(pathToRepoDocs);
+})();
